spring rest api

+++ ch08/src/main/java/kr/co/ch08/service/User1Service.java
@@ -15,8 +15,12 @@ public class User1Service {
 
     private final User1Repository user1Repository;
 
-    public void save(User1DTO user1DTO) {
-        user1Repository.save(user1DTO.toEntity());
+    public User1DTO save(User1DTO user1DTO) {
+
+        // Entity 저장 후 저장된 Entity 반환
+        User1 savedUser1 = user1Repository.save(user1DTO.toEntity());
+
+        return savedUser1.toDTO();
     }
     public User1DTO getUser(String userid){
 
@@ -36,12 +40,21 @@ public class User1Service {
                 .map(entity -> entity.toDTO())
                 .toList();
     }
-    public void modify(User1DTO user1DTO){
+    public User1DTO modify(User1DTO user1DTO){
         if(user1Repository.existsById(user1DTO.getUserid())){
-            user1Repository.save(user1DTO.toEntity());
+            User1 modifiedUser1 = user1Repository.save(user1DTO.toEntity());
+            return modifiedUser1.toDTO();
         }
+        return null;
     }
-    public void remove(String userid){
-        user1Repository.deleteById(userid);
+    public boolean remove(String userid){
+
+        if(user1Repository.existsById(userid)){
+            user1Repository.deleteById(userid);
+            return true;
+        }
+        return false;
     }
+
+
 }
